refactor: move test files over to the app-series
